Rename RootUtils -> root_utils
