+++ tests/test_rootfilereader.py
@@ -8,7 +8,7 @@ import ctypes
 import numpy as np
 import ROOT
 from test_utilities import float_compare, tuple_compare
-from hepdata_lib.RootUtils import RootFileReader, get_hist_2d_points
+from hepdata_lib.root_utils import RootFileReader, get_hist_2d_points
 
 
 class TestRootFileReader(TestCase):
